Add Utilities Lib

+++ EchoEngine/source/Actor.cpp
@@ -25,7 +25,7 @@ Actor::update(float deltaTime, DeviceContext deviceContext) {
 
 void 
 Actor::render(DeviceContext deviceContext) {
-	for (size_t i = 0; i < m_meshes.size(); i++) {
+	for (unsigned int i = 0; i < m_meshes.size(); i++) {
 		m_vertexBuffers[i].render(deviceContext, 0, 1);
 		m_indexBuffers[i].render(deviceContext, DXGI_FORMAT_R32_UINT);
 		if (m_textures.size() > 0)
@@ -40,7 +40,7 @@ Actor::render(DeviceContext deviceContext) {
 		}
 		m_sampler.render(deviceContext, 0, 1);
 		m_modelBuffer.renderModel(deviceContext, 2, 1);
-
+		
 		deviceContext.m_deviceContext->IASetPrimitiveTopology(D3D11_PRIMITIVE_TOPOLOGY_TRIANGLELIST);
 		deviceContext.m_deviceContext->DrawIndexed(m_meshes[i].m_numIndex, 0, 0);
 	}
